Renaming and deleting code that is not being used
