+++ node/events/buildStatus.ts
@@ -1,105 +1,122 @@
-import * as crypto from 'crypto'
+import { createHash } from 'crypto'
 
-import { VBase } from '@vtex/api'
+import { EventContext, VBase } from '@vtex/api'
 
-import { parseEvent, saveBuildInfo, validEvent } from '.'
-import errorHandler from './errorHandler'
+import { Clients } from '../clients'
+import { STORE_STATE } from '../util/constants'
+import { getBuildStatus, saveBuildStatus } from '../util/vbase'
 
-const buildCodeToGithubStatus: IStringToString = {
-  fail: 'failure',
-  success: 'success',
-}
-
-function validBuildCode(buildCode: string): boolean {
-  return buildCode in buildCodeToGithubStatus
-}
-
-export async function savePublishInfo(
-  ctx: ColossusEventContext,
-  vbase: VBase,
-  event: ColossusEvent
-): Promise<void> {
-  const { appId } = event
-  const [app, appVersion] = appId.split('@')
-
-  const publishExistencePath = `publish/${app}/${appVersion}/${event.buildId}`
-
-  vbase
-    .saveJSON('logs', publishExistencePath, {})
-    .catch(error => errorHandler(error, true, ctx.clients.logger, ctx))
-
-  const appExistencePath = `app/${app}`
-
-  vbase
-    .saveJSON('logs', appExistencePath, {})
-    .catch(error => errorHandler(error, true, ctx.clients.logger, ctx))
-}
-
-async function saveBuildStatusAsLog(
-  ctx: ColossusEventContext,
-  vbase: VBase,
-  event: ColossusEvent
-): Promise<void> {
-  let message = `build.status: ${event.buildCode}`
-
-  if (event.message) {
-    message += ` - ${event.message}`
-  }
-
-  const level = event.buildCode === 'success' ? 'info' : 'error'
-
-  const log = {
-    body: {
-      level,
-      message,
-      subject: 'subject',
-    },
-    key: event.key,
-    sender: event.sender,
-  }
-
-  const eventId = crypto.randomBytes(8).toString('hex')
-  const logPath = `build/logs/${event.buildId}/${eventId}`
-
-  vbase
-    .saveJSON('logs', logPath, log)
-    .catch(error => errorHandler(error, true, ctx.clients.logger, ctx))
-}
-
-async function saveBuildStatus(
-  ctx: ColossusEventContext,
-  vbase: VBase,
-  buildCode: string,
+export interface BuildStatus {
+  appId: string
+  buildCode: string
   buildId: string
-): Promise<void> {
-  const path = `build/status/${buildId}`
-
-  vbase
-    .saveJSON('logs', path, { buildCode })
-    .catch(error => errorHandler(error, true, ctx.clients.logger, ctx))
+  message: string
 }
 
-export default async function buildStatus(
-  ctx: ColossusEventContext
-): Promise<void> {
+export async function buildStatus(
+  ctx: EventContext<Clients, State>,
+  next: () => Promise<void>
+) {
   const event = parseEvent(ctx)
 
-  if (validEvent(event) && validBuildCode(event.buildCode)) {
-    const vbase = new VBase(ctx)
-
-    saveBuildStatus(ctx, vbase, event.buildCode, event.buildId).catch(error =>
-      errorHandler(error, true, ctx.clients.logger, ctx)
-    )
-    saveBuildStatusAsLog(ctx, vbase, event).catch(error =>
-      errorHandler(error, true, ctx.clients.logger, ctx)
-    )
-    saveBuildInfo(ctx, vbase, event).catch(error =>
-      errorHandler(error, true, ctx.clients.logger, ctx)
-    )
-    if (event.routeId === 'publish') {
-      savePublishInfo(ctx, vbase, event).catch(error =>
-        errorHandler(error, true, ctx.clients.logger, ctx)
-      )
-    }
+  if (!event.appId.includes(STORE_STATE)) {
+    return
   }
+
+  const { vbase } = ctx.clients
+
+  const previousStatus = (await getBuildStatus(
+    vbase,
+    ctx.vtex.account,
+    ctx.vtex.workspace
+  )) as BuildStatus
+
+  const buildId = `${ctx.vtex.account}.${ctx.vtex.workspace}`
+  const buildHash = createHash('md5')
+    .update(buildId)
+    .digest('hex')
+
+  if (buildHash !== previousStatus.buildId) {
+    return
+  }
+
+  if (event.buildCode === 'success') {
+    buildSuccess({
+      account: ctx.vtex.account,
+      previousStatus,
+      vbase,
+      workspace: ctx.vtex.workspace,
+    })
+  }
+
+  if (event.buildCode === 'fail') {
+    buildFail({
+      account: ctx.vtex.account,
+      event,
+      previousStatus,
+      vbase,
+      workspace: ctx.vtex.workspace,
+    })
+  }
+
+  next()
+}
+
+function parseEvent(ctx: EventContext<Clients, State>) {
+  const [senderName] = ctx.sender.split('@')
+  const { key, sender } = ctx
+  const { code, subject, message } = ctx.body
+  const { routeId } = ctx.body.builderHubIncomingRequestInfo
+  const { buildHash, trigger } = ctx.body.builderHubIncomingRequestInfo
+    .queryString || { trigger: 'toolbelt', ciHubRequestId: '' }
+
+  return {
+    appId: subject,
+    buildCode: code,
+    buildId: buildHash,
+    key,
+    message,
+    routeId,
+    sender,
+    senderName,
+    trigger,
+  } as ColossusEvent
+}
+
+interface BuildSuccessArgs {
+  previousStatus: BuildStatus
+  vbase: VBase
+  account: string
+  workspace: string
+}
+
+function buildSuccess({
+  previousStatus,
+  vbase,
+  account,
+  workspace,
+}: BuildSuccessArgs) {
+  previousStatus.buildCode = 'SUCCESS'
+  saveBuildStatus({ vbase, buildStatus: previousStatus, account, workspace })
+}
+
+interface BuildFailArgs {
+  event: ColossusEvent
+  previousStatus: BuildStatus
+  vbase: VBase
+  account: string
+  workspace: string
+}
+
+function buildFail({
+  event,
+  previousStatus,
+  vbase,
+  account,
+  workspace,
+}: BuildFailArgs) {
+  previousStatus.buildCode = 'BUILD_FAILED'
+  previousStatus.message = event.message
+
+  saveBuildStatus({ vbase, buildStatus: previousStatus, account, workspace })
 }
